Add integration test for get pack
