+++ server/service/packs_test.go
@@ -1,0 +1,30 @@
+package service
+
+import (
+	"context"
+	"testing"
+
+	"github.com/fleetdm/fleet/v4/server/datastore/mysql"
+	"github.com/fleetdm/fleet/v4/server/fleet"
+	"github.com/fleetdm/fleet/v4/server/test"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestGetPack(t *testing.T) {
+	ds := mysql.CreateMySQLDS(t)
+	defer ds.Close()
+
+	svc := newTestService(ds, nil, nil)
+
+	pack := &fleet.Pack{
+		Name: "foo",
+	}
+	_, err := ds.NewPack(context.Background(), pack)
+	assert.Nil(t, err)
+	assert.NotZero(t, pack.ID)
+
+	packVerify, err := svc.GetPack(test.UserContext(test.UserAdmin), pack.ID)
+	assert.Nil(t, err)
+
+	assert.Equal(t, pack.ID, packVerify.ID)
+}
